#359 move pandas_tool to brick module where it makes most sense.
