+++ src/f09_filter/test/test_bridgekind_dict.py
@@ -1,13 +1,13 @@
-from src.f00_instrument.pandas_tool import (
-    get_ordered_csv,
-    get_sorting_priority_column_headers as sorting_columns,
-)
 from src.f01_road.road import default_road_delimiter_if_none
 from src.f04_gift.atom_config import (
     type_AcctID_str,
     type_RoadNode_str,
     type_RoadUnit_str,
     type_GroupID_str,
+)
+from src.f08_brick.pandas_tool import (
+    get_ordered_csv,
+    get_sorting_priority_column_headers as sorting_columns,
 )
 from src.f09_filter.filter import (
     bridgekind_shop,
